adds support to pandas>=0.20
