+++ sparsity/test/test_dask_sparse_frame.py
@@ -1,3 +1,4 @@
+import datetime as dt
 import os
 import shutil
 import tempfile
@@ -8,9 +9,11 @@ import dask.dataframe as dd
 import numpy as np
 import pandas as pd
 import pytest
+from dask.async import get_sync
 
 import sparsity as sp
 import sparsity.dask as dsp
+from sparsity import sparse_one_hot
 from sparsity.dask.reshape import one_hot_encode
 
 dask.context.set_options(get=dask.async.get_sync)
@@ -67,6 +70,34 @@ def test_loc(iindexer, correct_shape):
     assert isinstance(res, sp.SparseFrame)
     assert res.shape == correct_shape
 
+def test_dask_loc(clickstream):
+    sf = dd.from_pandas(clickstream, npartitions=10) \
+        .map_partitions(
+        sparse_one_hot,
+        column='page_id',
+        categories=list('ABCDE'),
+        meta=list
+    )
+
+    res = sf.loc['2016-01-15':'2016-02-15']
+    res = sp.SparseFrame.concat(res.compute(get=get_sync).tolist())
+    assert res.index.date.max() == dt.date(2016, 2, 15)
+    assert res.index.date.min() == dt.date(2016, 1, 15)
+
+
+def test_dask_multi_index_loc(clickstream):
+    sf = dd.from_pandas(clickstream, npartitions=10) \
+        .map_partitions(
+            sparse_one_hot,
+            column='page_id',
+            index_col=['index', 'id'],
+            categories=list('ABCDE'),
+            meta=list
+    )
+    res = sf.loc['2016-01-15':'2016-02-15']
+    res = sp.SparseFrame.vstack(res.compute(get=get_sync).tolist())
+    assert res.index.get_level_values(0).date.min() == dt.date(2016, 1, 15)
+    assert res.index.get_level_values(0).date.max() == dt.date(2016, 2, 15)
 
 def test_repr():
     dsf = dsp.from_pandas(pd.DataFrame(np.random.rand(10, 2)),
